requirements.txt
streamlit 
langchain-openai 
python-dotenv 
pillow 
python-docx 
PyPDF2
